pyproject.toml
[tool.poetry]
authors = ["m09 <142691+m09@users.noreply.github.com>"]
description = ""
name = "lkdlt"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
genanki = "^0.13.0"
python = "^3.9"
rich = "^13.3.2"
typer = "^0.7.0"
selenium = "^4.8.2"
appdirs = "^1.4.4"
pydantic = "^1.10.6"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
flake8-bugbear = "^23.3.12"
isort = "^5.12.0"
mypy = "^1.1.1"
pylint = "^2.17.0"
pytest = "^7.2.2"
types-pyyaml = "^6.0.12.8"

[tool.poetry.scripts]
lkdlt = "lkdlt.cli:main"

[tool.isort]
profile = "black"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
